+++ src/i8x32_.rs
@@ -306,7 +306,7 @@ impl i8x32 {
   pub fn move_mask(self) -> u32 {
     pick! {
       if #[cfg(target_feature="avx2")] {
-        move_mask_i8_m256i(self.avx)
+        move_mask_i8_m256i(self.avx) as u32
       } else {
         self.a.move_mask() | (self.b.move_mask() << 16)
       }
