+++ src/u8x32_.rs
@@ -249,7 +249,7 @@ impl u8x32 {
   #[inline]
   #[must_use]
   pub fn move_mask(self) -> u32 {
-    i8x32::move_mask(cast(self))
+    i8x32::move_mask(cast(self)) as u32
   }
 
   #[inline]
